fix(PKH Files Upload): add feature flag and extract dokumente assertion to a new file
